Add init noist gaussion smooth
